Homework 4/requirements_issuer.txt
Flask==2.3.2
flask_cors
mysql-connector-python==8.0.33
